flask-board-app/requirements.txt
beautifulsoup4==4.10.0
black==21.9b0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
flake8==3.9.2
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-WTF==0.15.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.7.0
pyflakes==2.3.1
pymongo==3.12.0
python-dotenv==0.19.2
regex==2021.9.30
requests==2.26.0
soupsieve==2.3.1
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
